fix(triggers): polish Add trigger modal and catalog layout
- Move Copy button into editor corner so the Source tab has no empty row
- Render full markdown description in a hover tooltip on catalog cards
- Collapse TopNavBar empty space so Add/Manage tabs sit flush under title
- Revert over-broad Documentation tab background override that broke light mode
