+++ ui/src/components/admin/TriggersAdd.vue
@@ -1,27 +1,27 @@
 <template>
-    <section class="triggers-add container">
+    <div class="triggers-add">
         <div class="toolbar">
-            <el-input
-                v-model="searchQuery"
-                :placeholder="$t('triggers.add.search_placeholder')"
-                :prefix-icon="Magnify"
-                clearable
-                class="search-input"
-            />
-            <el-radio-group v-model="activeCategoryFilter" class="category-filter">
-                <el-radio-button label="all">
-                    {{ $t("triggers.add.filter.all") }}
-                </el-radio-button>
-                <el-radio-button label="core">
-                    {{ $t("triggers.add.filter.core") }}
-                </el-radio-button>
-                <el-radio-button label="realtime">
-                    {{ $t("triggers.add.filter.realtime") }}
-                </el-radio-button>
-                <el-radio-button label="app">
-                    {{ $t("triggers.add.filter.app") }}
-                </el-radio-button>
-            </el-radio-group>
+            <div class="search-wrapper">
+                <Magnify class="search-icon" />
+                <input
+                    v-model="searchQuery"
+                    type="text"
+                    class="search-input"
+                    :placeholder="$t('triggers.add.search_placeholder')"
+                >
+            </div>
+            <div class="filter-group">
+                <button
+                    v-for="option in filterOptions"
+                    :key="option.value"
+                    type="button"
+                    class="filter-button"
+                    :class="{active: activeCategoryFilter === option.value}"
+                    @click="activeCategoryFilter = option.value"
+                >
+                    {{ option.label }}
+                </button>
+            </div>
         </div>
 
         <div v-if="loading" class="state-empty">
@@ -35,16 +35,16 @@
 
         <template v-else>
             <TriggersCategorySection
-                v-if="activeCategoryFilter === 'all' || activeCategoryFilter === 'core'"
+                v-if="showCategory('core')"
                 category="core"
                 :title="$t('triggers.add.category.core.title')"
                 :description="$t('triggers.add.category.core.description')"
                 :triggers="coreTriggers"
-                :expand-all="true"
+                :expandAll="true"
                 @add="openConfigureModal"
             />
             <TriggersCategorySection
-                v-if="activeCategoryFilter === 'all' || activeCategoryFilter === 'realtime'"
+                v-if="showCategory('realtime')"
                 category="realtime"
                 :title="$t('triggers.add.category.realtime.title')"
                 :description="$t('triggers.add.category.realtime.description')"
@@ -52,7 +52,7 @@
                 @add="openConfigureModal"
             />
             <TriggersCategorySection
-                v-if="activeCategoryFilter === 'all' || activeCategoryFilter === 'app'"
+                v-if="showCategory('app')"
                 category="app"
                 :title="$t('triggers.add.category.app.title')"
                 :description="$t('triggers.add.category.app.description')"
@@ -67,11 +67,12 @@
             :trigger="selectedTrigger"
             @cancel="configureModalVisible = false"
         />
-    </section>
+    </div>
 </template>
 
 <script setup lang="ts">
     import {computed, onMounted, ref} from "vue";
+    import {useI18n} from "vue-i18n";
     import Magnify from "vue-material-design-icons/Magnify.vue";
 
     import {usePluginsStore, type TriggerPluginDto} from "../../stores/plugins";
@@ -80,6 +81,7 @@
 
     type CategoryFilter = "all" | "core" | "realtime" | "app";
 
+    const {t} = useI18n({useScope: "global"});
     const pluginsStore = usePluginsStore();
 
     const loading = ref(true);
@@ -91,19 +93,25 @@
 
     const MCP_TOOL_TYPE = "io.kestra.core.models.triggers.McpTool";
 
+    const filterOptions = computed<{value: CategoryFilter; label: string}[]>(() => [
+        {value: "all", label: t("triggers.add.filter.all")},
+        {value: "core", label: t("triggers.add.filter.core")},
+        {value: "realtime", label: t("triggers.add.filter.realtime")},
+        {value: "app", label: t("triggers.add.filter.app")},
+    ]);
+
     function filterBySearch(triggers: TriggerPluginDto[]) {
         const q = searchQuery.value.trim().toLowerCase();
         if (!q) return triggers;
-        return triggers.filter(t =>
-            t.name.toLowerCase().includes(q) ||
-            t.type.toLowerCase().includes(q) ||
-            (t.description ?? "").toLowerCase().includes(q)
+        return triggers.filter(tr =>
+            tr.name.toLowerCase().includes(q) ||
+            tr.type.toLowerCase().includes(q) ||
+            (tr.description ?? "").toLowerCase().includes(q)
         );
     }
 
     const coreTriggers = computed(() => {
-        const list = filterBySearch(allTriggers.value.filter(t => t.group === "core"));
-        // Pin MCP Tool trigger to the top when it is available.
+        const list = filterBySearch(allTriggers.value.filter(tr => tr.group === "core"));
         return [...list].sort((a, b) => {
             if (a.type === MCP_TOOL_TYPE) return -1;
             if (b.type === MCP_TOOL_TYPE) return 1;
@@ -112,12 +120,16 @@
     });
 
     const realtimeTriggers = computed(() =>
-        filterBySearch(allTriggers.value.filter(t => t.group === "realtime"))
+        filterBySearch(allTriggers.value.filter(tr => tr.group === "realtime"))
     );
 
     const appTriggers = computed(() =>
-        filterBySearch(allTriggers.value.filter(t => t.group === "app"))
+        filterBySearch(allTriggers.value.filter(tr => tr.group === "app"))
     );
+
+    function showCategory(group: "core" | "realtime" | "app"): boolean {
+        return activeCategoryFilter.value === "all" || activeCategoryFilter.value === group;
+    }
 
     const hasAnyVisibleTrigger = computed(() =>
         coreTriggers.value.length > 0 ||
@@ -145,32 +157,96 @@
 
 <style scoped lang="scss">
     .triggers-add {
-        padding: 1.5rem;
+        display: flex;
+        flex-direction: column;
+        gap: 18px;
     }
 
     .toolbar {
         display: flex;
-        gap: 1rem;
-        margin-bottom: 1.5rem;
+        gap: 12px;
         align-items: center;
         flex-wrap: wrap;
     }
 
-    .search-input {
-        max-width: 420px;
+    .search-wrapper {
+        position: relative;
         flex: 1 1 280px;
+        max-width: 520px;
     }
 
-    .category-filter {
-        flex-shrink: 0;
+    .search-icon {
+        position: absolute;
+        left: 0.75rem;
+        top: 50%;
+        transform: translateY(-50%);
+        color: var(--ks-content-tertiary, #888);
+        pointer-events: none;
+        font-size: 1rem;
+    }
+
+    .search-input {
+        width: 100%;
+        padding: 0.5rem 0.75rem 0.5rem 2.25rem;
+        background: var(--ks-background-input, var(--bs-body-bg));
+        border: 1px solid var(--ks-border-primary, var(--bs-border-color));
+        border-radius: 6px;
+        color: var(--ks-content-primary, var(--bs-body-color));
+        font-size: 0.875rem;
+
+        &:focus {
+            outline: none;
+            border-color: var(--el-color-primary);
+            box-shadow: 0 0 0 2px rgba(124, 58, 237, 0.15);
+        }
+
+        &::placeholder {
+            color: var(--ks-content-tertiary, #888);
+        }
+    }
+
+    .filter-group {
+        display: inline-flex;
+        border: 1px solid var(--ks-border-primary, var(--bs-border-color));
+        border-radius: 6px;
+        overflow: hidden;
+    }
+
+    .filter-button {
+        padding: 0.5rem 1rem;
+        background: transparent;
+        border: none;
+        color: var(--ks-content-secondary, var(--bs-gray-600));
+        font-size: 0.875rem;
+        font-weight: 500;
+        cursor: pointer;
+        transition: background-color 0.12s ease, color 0.12s ease;
+
+        & + .filter-button {
+            border-left: 1px solid var(--ks-border-primary, var(--bs-border-color));
+        }
+
+        &:hover {
+            color: var(--ks-content-primary, var(--bs-body-color));
+        }
+
+        &.active {
+            background: var(--el-color-primary);
+            color: #fff;
+        }
     }
 
     .state-empty {
-        padding: 2rem;
+        padding: 3rem 1rem;
         text-align: center;
 
         h4 {
-            margin-bottom: .5rem;
+            margin-bottom: 0.5rem;
+        }
+
+        p {
+            color: var(--ks-content-secondary, var(--bs-gray-600));
+            margin: 0;
         }
     }
 </style>
