+++ ui/src/styles/app.scss
@@ -257,3 +257,39 @@ body{
     left: 0;
     z-index: 10000;
 }
+
+// Trigger catalog card hover tooltip. Must be global because el-tooltip
+// teleports its popper out of the component tree, so scoped styles never
+// reach it.
+.el-popper.trigger-card-tooltip {
+    max-width: 420px;
+    font-size: 12px;
+    line-height: 1.5;
+    padding: 10px 12px;
+
+    p { margin: 0 0 6px; }
+    p:last-child { margin-bottom: 0; }
+
+    code {
+        font-family: var(--font-monospace, monospace);
+        font-size: 0.92em;
+        padding: 1px 4px;
+        border-radius: 3px;
+        background: rgba(255, 255, 255, 0.12);
+    }
+
+    pre {
+        margin: 6px 0;
+        padding: 8px 10px;
+        border-radius: 4px;
+        background: rgba(0, 0, 0, 0.35);
+        overflow-x: auto;
+
+        code { background: transparent; padding: 0; }
+    }
+
+    ul, ol { margin: 4px 0 6px 18px; padding: 0; }
+    li { margin-bottom: 2px; }
+
+    a { color: var(--el-color-primary-light-5, #c4b5fd); }
+}
