+++ ui/src/translations/en.json
@@ -413,7 +413,8 @@
     "triggers.add.category.app.title": "App Triggers",
     "triggers.add.category.app.description": "Periodically check external resources for changes.",
     "triggers.add.card.add_to_flow": "Add to flow",
-    "triggers.add.modal.title": "Configure {name}",
+    "triggers.add.card.add": "Add",
+    "triggers.add.modal.title": "{name}",
     "triggers.add.modal.tab.form": "Form",
     "triggers.add.modal.tab.source": "Source",
     "triggers.add.modal.tab.documentation": "Documentation",
@@ -426,6 +427,7 @@
     "triggers.add.modal.add_button": "Add Trigger to Flow",
     "copied_to_clipboard": "Copied to clipboard",
     "copy_failed": "Failed to copy to clipboard",
+    "copied": "Copied",
     "trigger details": "Trigger details",
     "triggerflow disabled": "Trigger is disabled from flow definition",
     "last trigger date": "Last triggered date",
